rewrote the acceptPayment method in restaurant to also clear the table and reset patron's tab

+++ app/src/main/java/com/example/marta/mariosrestaurant/tables/Table.java
@@ -37,7 +37,7 @@ public class Table {
         table.remove(patron);
     }
 
-    public int totalValue() {
+    public double totalValue() {
         int total = 0;
         for (Patron patron : table) {
             total += patron.calculateTotalShare();
